# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/payments/cancel": {
            "post": {
                "description": "Handles the cancellation callback from the Payment Service Provider (PSP) when a deposit is cancelled.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Cancel a Deposit Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PSP API Key",
                        "name": "X-PSP-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Cancellation callback from PSP",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/psp.CancelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deposit cancelled successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/payments/confirm": {
            "post": {
                "description": "Handles the confirmation callback from the Payment Service Provider (PSP) after a successful deposit.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Confirm a Deposit Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PSP API Key",
                        "name": "X-PSP-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Confirmation callback from PSP",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/psp.PayInResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deposit confirmed successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/payments/deposit": {
            "post": {
                "description": "Creates a new PENDING transaction and redirects the user to the Payment Service Provider (PSP) for payment completion.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Initiate a Deposit Transaction",
                "parameters": [
                    {
                        "description": "Deposit initiation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DepositRequest"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to PSP payment page",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.DepositRequest": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "user_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "psp.CancelRequest": {
            "type": "object",
            "properties": {
                "transactionID": {
                    "type": "string"
                }
            }
        },
        "psp.PayInResponse": {
            "type": "object",
            "properties": {
                "redirectUrl": {
                    "type": "string"
                },
                "transactionID": {
                    "type": "string"
                }
            }
        }
    }
}